Note about increasing max_samples_per_send
As found in issue #11, tuning may be needed for remote writes, to allow larger batches of metrics to be sent to ClickHouse in fewer writes.

I picked 10000 as a "safe for everyone" value that matches the current
Prometheus per-shard capacity default, but allows the entire queue to
be sent with one write.